backend/requirements.txt
# Django and Django REST Framework
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1

# Database
mysqlclient==2.2.0

# CORS handling
django-cors-headers==4.3.1

# Environment variables
python-dotenv==1.0.0

# Password hashing (bcrypt alternative, Django uses its own)
argon2-cffi==23.1.0

# Google OAuth
google-auth==2.25.2
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0

# Production server
gunicorn==21.2.0
whitenoise==6.6.0

# Utils
pytz==2023.3
